Made updates to all the files
For the index page, changed the build button into an anchor tag and linked it to the templates page
For the templates page, changed the individual template li elements into forms that can be used to pass their values from page to page. This can be better achieved with either PHP or JS though.
For the resumeform page, styled it appropriately and tried to retrieve the template value from the previous templates page that led to it
For the contact-us page, made slight changes to the nav element. If we were using PHP, would have used includes for everything, then have a smart nav.php file
Made numerous changes to the style.css file.